+++ VAMP_NBP/VAMP_Backend/src/de/cebitec/vamp/databackend/connector/ReferenceConnector.java
@@ -38,7 +38,7 @@ public class ReferenceConnector {
             ResultSet rs = fetch.executeQuery();
 
             if(rs.next()){
-                Long id = rs.getLong(FieldNames.REF_GEN_ID);
+                int id = rs.getInt(FieldNames.REF_GEN_ID);
                 String name = rs.getString(FieldNames.REF_GEN_NAME);
                 String description = rs.getString(FieldNames.REF_GEN_DESCRIPTION);
                 String sequence = rs.getString(FieldNames.REF_GEN_SEQUENCE);
@@ -92,10 +92,10 @@ public class ReferenceConnector {
 
             ResultSet rs = fetch.executeQuery();
             while(rs.next()){
-                Long id = rs.getLong(FieldNames.TRACK_ID);
+                int id = rs.getInt(FieldNames.TRACK_ID);
                 String description = rs.getString(FieldNames.TRACK_DESCRIPTION);
                 Timestamp date = rs.getTimestamp(FieldNames.TRACK_TIMESTAMP);
-                Long refGenomeID = rs.getLong(FieldNames.TRACK_REFERENCE_ID);
+                int refGenomeID = rs.getInt(FieldNames.TRACK_REFERENCE_ID);
 //                Long runID = rs.getLong(FieldNames.TRACK_RUN);
                 list.add(new PersistantTrack(id, description, date, refGenomeID));//, runID));
             }
